+++ src/main/java/de/fh_zwickau/informatik/sensor/model/locations/Location.java
@@ -8,6 +8,59 @@
  */
 package de.fh_zwickau.informatik.sensor.model.locations;
 
+import org.apache.commons.lang3.builder.ToStringBuilder;
+
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * The {@link Location} represents a location. See also:
+ * http://docs.zwayhomeautomation.apiary.io/#
+ *
+ * @author Patrick Hecker - Initial contribution
+ */
 public class Location {
-    // TODO
+    @SerializedName("id")
+    private Integer mId;
+    @SerializedName("title")
+    private String mTitle;
+
+    /**
+     * Instantiate a location with default values.
+     */
+    public Location() {
+        mId = -1;
+        mTitle = "";
+    }
+
+    public Integer getId() {
+        if (mId == null) {
+            mId = -1;
+        }
+        return mId;
+    }
+
+    public void setId(Integer id) {
+        this.mId = id;
+    }
+
+    public String getTitle() {
+        if (mTitle == null) {
+            mTitle = "";
+        }
+        return mTitle;
+    }
+
+    public void setTitle(String title) {
+        this.mTitle = title;
+    }
+
+    /*
+     * (non-Javadoc)
+     *
+     * @see java.lang.Object#toString()
+     */
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this).append("id", mId).append("title", mTitle).toString();
+    }
 }